% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_before_fn}
\alias{get_before_fn}
\title{A helper method to get the pre-s3memoize function of a s3memoized function.}
\usage{
get_before_fn(fn)
}
\arguments{
\item{fn}{function. The s3memoized function to look for.}
}
\description{
A helper method to get the pre-s3memoize function of a s3memoized function.
}

