% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_s3path}
\alias{get_s3path}
\title{A helper method to fetch the name of the s3 path that a certain call is stored in.}
\usage{
get_s3path(fn, ...)
}
\arguments{
\item{fn}{function. The function to find the digest for.}

\item{...}{Arguments passed to the function that the digest is for.}
}
\description{
A helper method to fetch the name of the s3 path that a certain call is stored in.
}

