% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.s3memoized}
\alias{is.s3memoized}
\title{A helper method to determine if a function is already s3memoized.}
\usage{
is.s3memoized(fn)
}
\arguments{
\item{fn}{function. The function to test.}
}
\description{
A helper method to determine if a function is already s3memoized.
}

