% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.s3memoized_function}
\alias{print.s3memoized_function}
\title{Print batched functions as they once were.}
\usage{
\method{print}{s3memoized_function}(x, ...)
}
\arguments{
\item{x}{function. The function to print.}

\item{...}{additional arguments to pass to print.}
}
\description{
Print batched functions as they once were.
}

