% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3memoize-package.r, R/s3memoize.R
\docType{package}
\name{s3memoize}
\alias{s3memoize}
\alias{s3memoize-package}
\title{s3memoize.}
\usage{
s3memoize(fn)
}
\arguments{
\item{fn}{function. Function to s3memoize.}
}
\description{
s3memoize.

s3memoize maps a function to a s3memoized version of that function.
}

